pyproject.toml
[project]
name = "skillsyncer"
version = "0.1.0"
description = "Agent skills that sync, fill, and protect themselves."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "gitpython>=3.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
skillsyncer = "skillsyncer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["skillsyncer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
